+++ code/_global_vars/tgui_ch.dm
@@ -1,0 +1,1 @@
+GLOBAL_DATUM(crew_manifest_tgui, /datum/crew_manifest)
