+++ code/modules/admin/secrets/admin_secrets/show_crew_manifest.dm
@@ -5,8 +5,16 @@
 	. = ..()
 	if(!.)
 		return
+	/* CHOMPEdit Start - Crew Manifest
 	var/dat
 	dat += "<h4>Crew Manifest</h4>"
-	dat += data_core.get_manifest()
+	dat += data_core.get_manifest_html()
 
 	user << browse(dat, "window=manifest;size=370x420;can_close=1")
+	*/
+
+	if(!GLOB.crew_manifest_tgui)
+		GLOB.crew_manifest_tgui = new /datum/crew_manifest(src)
+
+	GLOB.crew_manifest_tgui.tgui_interact(src)
+	// CHOMPedit End
