+++ code/modules/mob/new_player/new_player.dm
@@ -730,13 +730,20 @@
 	return new_character
 
 /mob/new_player/proc/ViewManifest()
+	/* CHOMPEdit Start - Crew manifest
 	var/dat = "<div align='center'>"
-	dat += data_core.get_manifest(OOC = 1)
+	dat += data_core.get_manifest_html(OOC = 1)
 
 	//src << browse(dat, "window=manifest;size=370x420;can_close=1")
 	var/datum/browser/popup = new(src, "Crew Manifest", "Crew Manifest", 370, 420, src)
 	popup.set_content(dat)
 	popup.open()
+	*/
+	if(!GLOB.crew_manifest_tgui)
+		GLOB.crew_manifest_tgui = new /datum/crew_manifest(src)
+
+	GLOB.crew_manifest_tgui.tgui_interact(src)
+	// CHOMPEdit End
 
 /mob/new_player/Move()
 	return 0
