+++ vorestation.dme
@@ -124,6 +124,7 @@
 #include "code\_global_vars\misc.dm"
 #include "code\_global_vars\mobs.dm"
 #include "code\_global_vars\sensitive.dm"
+#include "code\_global_vars\tgui_ch.dm"
 #include "code\_global_vars\time_vars.dm"
 #include "code\_global_vars\traits.dm"
 #include "code\_global_vars\typecache.dm"
@@ -4688,6 +4689,7 @@
 #include "modular_chomp\code\modules\mining\shelters_ch.dm"
 #include "modular_chomp\code\modules\mob\holder.dm"
 #include "modular_chomp\code\modules\mob\mob.dm"
+#include "modular_chomp\code\modules\mob\dead\crew_manifest.dm"
 #include "modular_chomp\code\modules\mob\dead\observer\observer.dm"
 #include "modular_chomp\code\modules\mob\language\language.dm"
 #include "modular_chomp\code\modules\mob\living\emote.dm"
